Refactor: Implement Robot Pattern in NavigationTest using NavigationRobot.
